+++ lib/Service/Index.php
@@ -191,7 +191,7 @@ class Index
             $this->log("Indexing file {$path}", true);
             $this->timelineWrite->processFile($file);
         } catch (\OCP\Lock\LockedException $e) {
-            $this->log("Skipping file {$path} due to lock\n", true);
+            $this->log("Skipping file {$path} due to lock", true);
         } catch (\Exception $e) {
             $this->error("Failed to index file {$path}: {$e->getMessage()}");
         } finally {